GH-4: Add nginx
Added nginx which is used in both development and production
environments. Also moved all static files to the nginx container
exclusively. This is so that we have transparency that it *must be*
nginx that is serving the static files. Also this allows us to share
static files more easily across applications.