GH-9: Set up a postgres db
Set up a way for api-finances to communicate with a postgres
database. Created an end point that will automatically insert
a new instance of a model into the database. Created a way to
initalize the database and to drop the tables in the dabase,
they are as follows.

docker-compose run api-finances flask initdb
docker-compose run api-finances flask dropdb

TODO: Put this information in the README